various warnings

+++ shedskin/lib/array.hpp
@@ -291,14 +291,14 @@ template<> inline __ss_int array<__ss_int>::__getfast__(__ss_int i) {
 }
 template<> inline str *array<str *>::__getfast__(__ss_int i) {
     i = __wrap(this, i);
-    return __char_cache[(unsigned char)units[i]];
+    return __char_cache[(unsigned char)units[(size_t)i]];
 }
 template<> inline __ss_float array<__ss_float>::__getfast__(__ss_int i) {
     i = __wrap(this, i);
     if(typechar == 'f')
-        return *((float *)(&units[i*itemsize]));
+        return *((float *)(&units[(size_t)i*itemsize]));
     else
-        return *((__ss_float *)(&units[i*itemsize]));
+        return *((__ss_float *)(&units[(size_t)i*itemsize]));
 }
 
 template<class T> void *array<T>::append(T t) {
